exec: fix heredoc within pipes
It used to be that using a heredoc within a pipe had no result:

    TOSH $ ls | cat <<EOF
    TOSH $

This commit fixes that by moving the heredoc requesting outside of the fork
