+++ src/exec/redirect/handle_redirections.c
@@ -23,8 +23,7 @@ const struct s_redirection_kvp	g_redir_tbl[] = {
 };
 
 static t_error	redirect(t_list_meta *const tracker_lst,
-					const struct s_io_redirect *const redir,
-					const t_env *const env)
+					const struct s_io_redirect *const redir)
 {
 	int								first_fd;
 	const struct s_redirection_kvp	*redir_type;
@@ -45,7 +44,7 @@ static t_error	redirect(t_list_meta *const tracker_lst,
 			first_fd = 0;
 		if (exec__is_protected_fd(first_fd) == true)
 			return (errorf("%d is a protected fd", redir->fd));
-		return (redirect_heredoc(tracker_lst, first_fd, redir->here, env));
+		return (redirect_heredoc(tracker_lst, first_fd, redir->here));
 	}
 }
 
@@ -67,7 +66,7 @@ static t_error	recurse_prefix(t_list_meta *const tracker_lst,
 			return (err);
 		}
 	}
-	err = redirect(tracker_lst, prefix->redirect, env);
+	err = redirect(tracker_lst, prefix->redirect);
 	return (err);
 }
 
@@ -110,7 +109,7 @@ t_error			exec__handle_redirections(
 	{
 		if (suffix->redirect)
 		{
-			err = redirect(tracker_lst, suffix->redirect, env);
+			err = redirect(tracker_lst, suffix->redirect);
 			if (is_error(err))
 			{
 				return (err);
