+++ src/parser/io_here.c
@@ -57,6 +57,7 @@ void						free_io_here(struct s_io_here *const io_here)
 	if (io_here)
 	{
 		free(io_here->here_end);
+		free(io_here->contents);
 		free(io_here);
 	}
 }
